change routes "Users" in hw05
