+++ app.js
@@ -6,7 +6,6 @@ const mongoose = require("mongoose");
 
 const authRouter = require("./routes/api/auth");
 const contactsRouter = require("./routes/api/contacts");
-const usersRouter = require("./routes/api/users");
 
 require("dotenv").config();
 const { DB_HOST } = process.env;
@@ -28,7 +27,6 @@ app.use(express.static("public"));
 
 app.use("/api/auth", authRouter);
 app.use("/api/contacts", contactsRouter);
-app.use("/users", usersRouter);
 
 app.use((req, res) => {
   res.status(404).json({ message: "Not found" });
